cdk/lambda/requirements.txt
aws-cdk.aws-s3
aws-cdk.aws-s3-deployment
aws-cdk.aws-lambda
aws-cdk.aws-iam